Make Factory tabs active
