+++ front/src/shared/containers/agentManager.jsx
@@ -7,275 +7,16 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import Zrmc, { Grid, Inner, Cell } from "zrmc";
+import { Grid, Inner, Cell } from "zrmc";
 import Loading from "zoapp-front/components/loading";
 import SignInForm from "zoapp-front/containers/signInForm";
-import { appSetTitleName } from "zoapp-front/actions/app";
 
-import {
-  apiGetIntentsRequest,
-  apiSendIntentRequest,
-  apiDeleteIntentRequest,
-  apiSaveBotRequest,
-  apiImportRequest,
-} from "../actions/api";
-import { appUpdateIntent } from "../actions/app";
 import ExplorerContainer from "./explorerContainer";
 import IntentContainer from "./builder/intentContainer";
 import EntityContainer from "./builder/entityContainer";
 import FunctionContainer from "./builder/functionContainer";
-import PlaygroundContainer from "./playgroundContainer";
-import IODialog from "./dialogs/ioDialog";
-import FileManager from "../utils/fileManager";
 
 class AgentManager extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { needUpdate: true };
-    this.props.appSetTitleName(props.bot ? props.bot.name : "Factory");
-  }
-
-  componentDidMount() {
-    this.update();
-  }
-
-  componentDidUpdate() {
-    this.update();
-  }
-
-  update() {
-    if (this.updateIntents()) {
-      this.props.appSetTitleName(
-        this.props.bot ? this.props.bot.name : "Chatbot",
-        "Factory",
-      );
-    }
-  }
-
-  onRenameIntent = (dialog, action, data) => {
-    if (action === "Rename") {
-      const intentName = dialog.getFieldValue();
-      // console.log("WIP", `ExplorerContainer.onRenameIntent :${intentName}`);
-      if (intentName === "") {
-        dialog.invalidateField();
-        return false;
-      }
-      const { selected } = data;
-      const it = this.props.intents[selected];
-      const intent = { ...it, name: intentName };
-      this.props.apiSendIntentRequest(this.props.selectedBotId, intent);
-    }
-    return true;
-  };
-
-  addIntent(intentName, data) {
-    let { input } = data;
-    if (!input) {
-      input = [intentName];
-    }
-    let { output } = data;
-    if (!output) {
-      // TODO create a better output
-      let i = input[0];
-      const il = i.toLowerCase();
-      if (!(il === "hello" || il === "hi")) {
-        i = "I don't understand.";
-      }
-      output = [i];
-    }
-    const intent = {
-      input,
-      output,
-      name: intentName,
-    };
-    this.props.apiSendIntentRequest(this.props.selectedBotId, intent);
-  }
-
-  onAddIntent = (dialog, action, data) => {
-    if (action === "Create") {
-      const intentName = dialog.getFieldValue();
-      this.addIntent(intentName, data);
-    }
-    return true;
-  };
-
-  onDeleteIntent = (dialog, action, data) => {
-    if (action === "Delete") {
-      const { selected } = data;
-      const intent = this.props.intents[selected];
-      // console.log("WIP", `ExplorerContainer.onDeleteIntent :${intent.name}`);
-      this.props.apiDeleteIntentRequest(this.props.selectedBotId, intent);
-    }
-    return true;
-  };
-
-  onImportData = (data, options) => {
-    // console.log("AgentManager.onUpload=", options.filetype);
-    if (
-      options.filetype === "application/json" ||
-      options.filetype === "text/csv"
-    ) {
-      // WIP detect format
-      // console.log("AgentManager.onUpload=", data);
-      this.props.apiImportRequest(this.props.selectedBotId, data, options);
-    }
-  };
-
-  onDownloadData = () => {
-    const { name } = this.props.bot;
-    const data = { name, intents: this.props.intents };
-    const json = JSON.stringify(data);
-    FileManager.download(json, `${name}.json`, "application/json,.csv", () => {
-      /* console.log("ExplorerContainer.onDownload=", name); */
-    });
-  };
-
-  onEditWelcome = (dialog, action) => {
-    if (action === "Save") {
-      const welcome = dialog.getFieldValue();
-      if (welcome === "") {
-        dialog.invalidateField();
-        return false;
-      }
-      const bot = { ...this.props.bot, welcome };
-      this.props.apiSaveBotRequest(bot);
-    }
-    return true;
-  };
-
-  handleDeleteIntent = (selected = this.props.selectedIntentIndex) => {
-    // const selected = this.props.selectedIntentIndex;
-    const intent = this.props.intents[selected];
-    Zrmc.showDialog({
-      header: "Intent",
-      body: `${intent.name} Do you want to delete it ?`,
-      actions: [{ name: "Cancel" }, { name: "Delete" }],
-      onAction: this.onDeleteIntent,
-      data: { selected },
-    });
-  };
-
-  findIntentName(name) {
-    const { intents } = this.props;
-
-    let result = false;
-    if (intents && intents.length > 0) {
-      intents.forEach((intent) => {
-        if (intent.name.toLowerCase() === name) {
-          result = true;
-        }
-      });
-    }
-    return result;
-  }
-
-  generateIntentName() {
-    let name = "Intent";
-    let index = 1;
-    while (this.findIntentName(name.toLowerCase())) {
-      name = `Intent ${index}`;
-      index += 1;
-    }
-    return name;
-  }
-
-  handleAddIntent = (defaultValue = this.generateIntentName(), data = {}) => {
-    /* const field = {
-      defaultValue,
-      pattern: ".+",
-      name: "Intent name",
-      error: "Wrong name",
-    };
-    Zrmc.showDialog({
-      header: "Add new intent",
-      field,
-      actions: [{ name: "Cancel" }, { name: "Create" }],
-      onAction: this.onAddIntent,
-      data,
-    }); */
-    this.addIntent(defaultValue, data);
-  };
-
-  handleRenameIntent = (selected = this.props.selectedIntentIndex) => {
-    const intent = this.props.intents[selected];
-    const field = {
-      defaultValue: intent.name,
-      pattern: ".+",
-      name: "Intent name",
-      error: "Wrong name",
-    };
-    Zrmc.showDialog({
-      header: "Rename intent",
-      field,
-      actions: [{ name: "Cancel" }, { name: "Rename" }],
-      onAction: this.onRenameIntent,
-      data: { selected },
-    });
-  };
-
-  handleExportImport = (importOnly = false) => {
-    const dialog = (
-      <IODialog
-        open
-        importOnly={importOnly}
-        store={this.props.store}
-        onClosed={this.handleCloseDialog}
-        accept="application/json"
-        onDownload={this.onDownloadData}
-        onImport={this.onImportData}
-      />
-    );
-    Zrmc.showDialog(dialog);
-  };
-
-  handlePlaygroundAction = (action, defaultValue = "", data = {}) => {
-    if (action === "welcomeMessage") {
-      const field = {
-        defaultValue,
-        pattern: ".+",
-        name: "Welcome message",
-        error: "Wrong message",
-      };
-      Zrmc.showDialog({
-        header: "Edit Bot",
-        field,
-        actions: [{ name: "Cancel" }, { name: "Save" }],
-        onAction: this.onEditWelcome,
-        data,
-      });
-    } else if (action === "createIntent") {
-      const intent = { input: [defaultValue] };
-      this.handleAddIntent(defaultValue, intent);
-    } else if (action === "addInput") {
-      const intent = { ...this.props.selectedIntent };
-      if (!intent.input) {
-        intent.input = [];
-      }
-      // Check if sentence already exists in input
-      if (!intent.input.find((i) => i === defaultValue)) {
-        intent.input.push(defaultValue);
-        // this.props.appUpdateIntent(this.props.selectedBotId, intent);
-        delete intent.notSaved;
-        this.props.apiSendIntentRequest(this.props.selectedBotId, intent);
-      }
-    }
-    // console.log("botManager.handlePlaygroundAction", action);
-  };
-
-  updateIntents() {
-    if (!this.props.isSignedIn) {
-      if (!this.state.needUpdate) {
-        this.setState({ needUpdate: true });
-      }
-    }
-    if (this.props.selectedBotId && this.state.needUpdate) {
-      this.setState({ needUpdate: false });
-      this.props.apiGetIntentsRequest(this.props.selectedBotId);
-      return true;
-    }
-    return false;
-  }
-
   render() {
     let { isLoading } = this.props;
     if (!isLoading && !this.props.intents && this.props.isSignedIn) {
@@ -293,13 +34,12 @@ class AgentManager extends Component {
       <Cell
         style={{ margin: "0px", backgroundColor: "#f2f2f2" }}
         className="zui-color--white zui-panel"
-        span={2}
+        span={3}
       >
         <ExplorerContainer
-          handleExportImport={this.handleExportImport}
-          handleRename={this.handleRenameIntent}
-          handleAdd={this.handleAddIntent}
-          handleDelete={this.handleDeleteIntent}
+          onRename={this.props.onRenameIntent}
+          onAdd={this.props.onAddIntent}
+          onDelete={this.props.onDeleteIntent}
         />
       </Cell>
     );
@@ -308,7 +48,7 @@ class AgentManager extends Component {
         <Cell
           style={{ margin: "0px", backgroundColor: "#f2f2f2" }}
           className="zui-color--white zui-panel"
-          span={6}
+          span={9}
         >
           <FunctionContainer />
         </Cell>
@@ -318,7 +58,7 @@ class AgentManager extends Component {
         <Cell
           style={{ margin: "0px", backgroundColor: "#f2f2f2" }}
           className="zui-color--white zui-panel"
-          span={6}
+          span={9}
         >
           <EntityContainer />
         </Cell>
@@ -328,51 +68,22 @@ class AgentManager extends Component {
         <Cell
           style={{ margin: "0px", backgroundColor: "#f2f2f2" }}
           className="zui-color--white zui-panel"
-          span={6}
+          span={9}
         >
           <IntentContainer handleRename={this.handleRenameIntent} />
         </Cell>
       );
     }
-    const intentsEx = [];
-    if (Array.isArray(this.props.intents)) {
-      this.props.intents.forEach((intent) => {
-        const selected =
-          this.props.selectedIntent &&
-          intent.id === this.props.selectedIntent.id;
-        intentsEx.push({
-          id: intent.id,
-          selected,
-          name: intent.name,
-        });
-      });
-    }
     return (
-      <div className="zui-color--grey-100">
-        <Grid
-          gutter={{ desktop: "0px", tablet: "0px", phone: "0px" }}
-          style={{ margin: "0px", padding: "0px" }}
-        >
-          <Inner style={{ gridGap: "0px" }}>
-            {panel1}
-            {panel2}
-            <Cell
-              style={{
-                margin: "0px",
-                backgroundColor: "#f2f2f2",
-                zIndex: "3",
-              }}
-              className="mdc-elevation--z4"
-              span={4}
-            >
-              <PlaygroundContainer
-                onAction={this.handlePlaygroundAction}
-                intents={intentsEx}
-              />
-            </Cell>
-          </Inner>
-        </Grid>
-      </div>
+      <Grid
+        gutter={{ desktop: "0px", tablet: "0px", phone: "0px" }}
+        style={{ margin: "0px", padding: "0px" }}
+      >
+        <Inner style={{ gridGap: "0px" }}>
+          {panel1}
+          {panel2}
+        </Inner>
+      </Grid>
     );
   }
 }
@@ -402,13 +113,10 @@ AgentManager.propTypes = {
   selectedIntent: PropTypes.shape({ id: PropTypes.string }),
   store: PropTypes.shape({}),
   titleName: PropTypes.string.isRequired,
-  appSetTitleName: PropTypes.func.isRequired,
-  apiGetIntentsRequest: PropTypes.func.isRequired,
-  apiSendIntentRequest: PropTypes.func.isRequired,
-  apiDeleteIntentRequest: PropTypes.func.isRequired,
-  apiSaveBotRequest: PropTypes.func.isRequired,
-  apiImportRequest: PropTypes.func.isRequired,
-  appUpdateIntent: PropTypes.func.isRequired,
+
+  onRenameIntent: PropTypes.func.isRequired,
+  onAddIntent: PropTypes.func.isRequired,
+  onDeleteIntent: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = (state) => {
@@ -443,29 +151,7 @@ const mapStateToProps = (state) => {
   };
 };
 
-const mapDispatchToProps = (dispatch) => ({
-  apiGetIntentsRequest: (botId) => {
-    dispatch(apiGetIntentsRequest(botId));
-  },
-  appSetTitleName: (title, name) => {
-    dispatch(appSetTitleName(title, name));
-  },
-  apiSendIntentRequest: (botId, intent) => {
-    dispatch(apiSendIntentRequest(botId, intent));
-  },
-  apiDeleteIntentRequest: (botId, intentId) => {
-    dispatch(apiDeleteIntentRequest(botId, intentId));
-  },
-  apiSaveBotRequest: (bot) => {
-    dispatch(apiSaveBotRequest(bot));
-  },
-  apiImportRequest: (botId, data, options) => {
-    dispatch(apiImportRequest(botId, data, options));
-  },
-  appUpdateIntent: (botId, intent) => {
-    dispatch(appUpdateIntent(botId, intent));
-  },
-});
+const mapDispatchToProps = () => ({});
 
 // prettier-ignore
 export default connect(
